requirements.txt
streamlit==1.24.0
google-generativeai==0.8.5
google-api-core
google-api-python-client
google-auth
protobuf
pydantic
tqdm
typing-extensions
